requirements.txt
python-telegram-bot
pyttsx3
configparser